[level 0] Title: 두 수의 나눗셈, Time: 0.01 ms, Memory: 4.28 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.cpp
@@ -4,5 +4,6 @@
 using namespace std;
 
 int solution(int num1, int num2) {
-    return ((double)num1/(double)num2) * 1000;
+    
+    return ((double)num1 / num2) * 1000;
 }
